+++ functions/runtime_test.go
@@ -69,16 +69,16 @@ func TestAllCases(t *testing.T) {
 
 				require.NoError(t, err)
 
-				// Check that the whole project, including generated code, typechecks
-				typecheckResult, output := typecheck(workingDir)
-
-				assert.True(t, typecheckResult, output)
-
 				// Bundle all of the generated typescript code in @teamkeel/client
 				// necessary to run the node server
 				_, errs := runtime.Bundle(true)
 
 				require.Len(t, errs, 0)
+
+				// Check that the whole project, including generated code, typechecks
+				typecheckResult, output := typecheck(workingDir)
+
+				assert.True(t, typecheckResult, output)
 
 				port := 3002
 
@@ -162,7 +162,7 @@ func RunServer(workingDir string, port int) (*os.Process, error) {
 		return nil, err
 	}
 
-	cmd := exec.Command("node", serverDistPath)
+	cmd := exec.Command("node", filepath.Join("node_modules", "@teamkeel", "client", "dist", "handler.js"))
 	cmd.Env = append(cmd.Env, fmt.Sprintf("PORT=%d", port))
 	cmd.Dir = workingDir