+++ functions/sdk/jest.config.js
@@ -1,0 +1,7 @@
+const config = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  verbose: true,
+};
+
+export default config;