+++ functions/sdk/src/index.ts
@@ -1,0 +1,11 @@
+import Query from './query';
+import * as Constraints from './constraints';
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+//@ts-ignore
+export * from '../../client';
+
+export {
+  Query,
+  Constraints
+};